+++ CHANGES.md
@@ -6,6 +6,7 @@ GH::watch changelog
 - \#94 added app setting option to enable background notification check over WiFi only to save mobile data
 - \#92 added longer period options (6h, 12h, 24h) to "New notifications check period" 
 - \#93 changed default value of "New notifications check period" preference from 5 minutes to 1 hour to lover network bandwidth
+- \#96 patched UI bug in "filter by repository" - list of repos scroll back to the top possible, correct dismiss of dialog when swiped down  
 
 ## v1.21 (Oct 28th, 2017)
 - \#46 update to the Material design